packages/gcp/pyproject.toml
[project]
name = "pragmatiks-gcp-provider"
version = "0.77.0"
description = "GCP provider for Pragmatiks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pragmatiks-sdk>=0.6.0",
    "google-cloud-container>=2.50.0",
    "google-cloud-logging>=3.10.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
]

authors = [
    { name = "Pragmatiks", email = "team@pragmatiks.io" },
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
]

[tool.pragma]
# Provider metadata - used by the platform for discovery and deployment
provider = "gcp"
package = "gcp_provider"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.77.0"
version_files = ["pyproject.toml:^version"]
tag_format = "gcp-v$version"
update_changelog_on_bump = true
changelog_file = "CHANGELOG.md"

[tool.ruff]
line-length = 120

[tool.uv.build-backend]
module-name = "gcp_provider"

[build-system]
requires = ["uv_build>=0.9.8,<0.10.0"]
build-backend = "uv_build"
